pyproject.toml
[tool.poetry]
name = "user"
version = "0.1.0"
description = "user api"
authors = ["Akiva Zonenfeld <akivazo1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
pymongo = "^4.9.1"
python-dotenv = "^1.0.1"
waitress = "^3.0.0"
pytest = "^8.3.3"
mongomock = "^4.2.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
